split out lambda section tests into traversal and efficiency
